DTL :: double MAX_DISTANCE instead of int - less castings

+++ app/src/main/java/com/worldventures/dreamtrips/modules/dtl/model/merchant/filter/DtlFilterData.java
@@ -29,7 +29,7 @@ public abstract class DtlFilterData {
     }
 
     @Value.Default
-    public int getMaxDistance() {
+    public double getMaxDistance() {
         return DtlFilterParameters.MAX_DISTANCE;
     }
